Fix overflow issue in large-ish experiments

+++ src/main/java/ca/ubc/bps/state/MonitoredMutableDouble.java
@@ -25,7 +25,7 @@ public class MonitoredMutableDouble implements MutableDouble
   
   public static class ModCount
   {
-    public int count = 0;
+    public long count = 0;
   }
   
 }
